updated menu functionality. need to redesign prefs and n-panel ui.
